pyproject.toml
[project]
name = "sikuliplus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "robotframework>=7.3.2",
    "robotframework-sikulilibrary>=2.0.5",
]

[tool.sikuliplus]
similarity = 0.7
timeout = 1.0
action_speed = 0.1
highlight = true
highlight_time = 2.0
language = "pt_BR"
